Move EmailURIUniversalCommunication to DataType
